third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/credentialless/cors-or-credentialless/cross-origin-isolated-expected.txt

This is a testharness.js-based test.
FAIL Main crossOriginIsolated case: assert_equals: expected "crossOriginIsolated" but got "not isolated"
PASS Missing HTTPS:
PASS Missing COOP:same-origin:
PASS Report-only:
FAIL Report-only + enforced: assert_equals: expected "crossOriginIsolated" but got "not isolated"
Harness: the test ran to completion.
